# NonBacktracking inner matching loop optimizations