--- a/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/SymbolicRegexBuilder.cs
+++ b/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/SymbolicRegexBuilder.cs
@@ -120,9 +120,11 @@ namespace System.Text.RegularExpressions.Symbolic
         /// via https://github.com/dotnet/runtime/issues/63474), and use of Volatile.Read is
         /// contributing non-trivial overhead (https://github.com/dotnet/runtime/issues/65789).
         /// </remarks>
-        internal DfaMatchingState<TSet>?[]? _delta;
+        internal int[]? _delta;
         internal List<(DfaMatchingState<TSet>, DerivativeEffect[])>?[]? _capturingDelta;
         private const int InitialStateLimit = 1024;
+
+        internal (bool IsInitial, bool IsDeadend, bool IsNullable, bool CanBeNullable)[]? _stateInfo;
 
         /// <summary>1 + Log2(_minterms.Length), the smallest k s.t. 2^k >= minterms.Length + 1</summary>
         internal int _mintermsLog;
@@ -173,8 +175,10 @@ namespace System.Text.RegularExpressions.Symbolic
 
                 // the extra +1 slot with id minterms.Length is reserved for \Z (last occurrence of \n)
                 _mintermsLog = BitOperations.Log2((uint)_minterms.Length) + 1;
-                _delta = new DfaMatchingState<TSet>[InitialStateLimit << _mintermsLog];
+                _delta = new int[InitialStateLimit << _mintermsLog];
                 _capturingDelta = new List<(DfaMatchingState<TSet>, DerivativeEffect[])>[InitialStateLimit << _mintermsLog];
+
+                _stateInfo = new (bool, bool, bool, bool)[InitialStateLimit];
             }
 
             // initialized to False but updated later to the actual condition ony if \b or \B occurs anywhere in the regex
@@ -208,10 +212,10 @@ namespace System.Text.RegularExpressions.Symbolic
         }
 
         /// <summary>Returns the span from <see cref="_delta"/> that may contain transitions for the given state</summary>
-        internal Span<DfaMatchingState<TSet>?> GetDeltasFor(DfaMatchingState<TSet> state)
+        internal Span<int> GetDeltasFor(DfaMatchingState<TSet> state)
         {
             if (_delta is null || _minterms is null)
-                return Span<DfaMatchingState<TSet>?>.Empty;
+                return Span<int>.Empty;
             int numMinterms = state.StartsWithLineAnchor ? _minterms.Length + 1 : _minterms.Length;
             return _delta.AsSpan(state.Id << _mintermsLog, numMinterms);
         }
@@ -453,8 +457,9 @@ namespace System.Text.RegularExpressions.Symbolic
         /// <param name="node">the pattern that this state will represent</param>
         /// <param name="prevCharKind">the kind of the character that led to this state</param>
         /// <param name="capturing">whether to use the separate space of states with capturing transitions or not</param>
+        /// <param name="isInitialState">whether to mark the state as an initial state or not</param>
         /// <returns></returns>
-        public DfaMatchingState<TSet> CreateState(SymbolicRegexNode<TSet> node, uint prevCharKind, bool capturing = false)
+        public DfaMatchingState<TSet> CreateState(SymbolicRegexNode<TSet> node, uint prevCharKind, bool capturing = false, bool isInitialState = false)
         {
             //first prune the anchors in the node
             TSet wlbSet = _wordLetterForBoundariesSet;
@@ -469,21 +474,21 @@ namespace System.Text.RegularExpressions.Symbolic
             var s = new DfaMatchingState<TSet>(pruned_node, prevCharKind);
             if (!(capturing ? _capturingStateCache : _stateCache).TryGetValue(s, out DfaMatchingState<TSet>? state))
             {
-                state = MakeNewState(s, capturing);
+                state = MakeNewState(s, capturing, isInitialState);
             }
 
             return state;
         }
 
-        private DfaMatchingState<TSet> MakeNewState(DfaMatchingState<TSet> state, bool capturing)
+        private DfaMatchingState<TSet> MakeNewState(DfaMatchingState<TSet> state, bool capturing, bool isInitialState)
         {
             lock (this)
             {
                 HashSet<DfaMatchingState<TSet>> cache = capturing ? _capturingStateCache : _stateCache;
+                cache.Add(state); // Add to cache first to make 1 the first state ID
                 state.Id = cache.Count;
-                cache.Add(state);
 
-                Debug.Assert(_stateArray is not null && _capturingStateArray is not null);
+                Debug.Assert(_stateArray is not null && _capturingStateArray is not null && _stateInfo is not null);
 
                 const int GrowthSize = 1024;
                 if (capturing)
@@ -503,8 +508,10 @@ namespace System.Text.RegularExpressions.Symbolic
                         int newsize = _stateArray.Length + GrowthSize;
                         Array.Resize(ref _stateArray, newsize);
                         Array.Resize(ref _delta, newsize << _mintermsLog);
+                        Array.Resize(ref _stateInfo, newsize);
                     }
                     _stateArray[state.Id] = state;
+                    _stateInfo[state.Id] = (isInitialState, state.IsDeadend, state.IsNullable, state.CanBeNullable);
                 }
                 return state;
             }
@@ -549,13 +556,20 @@ namespace System.Text.RegularExpressions.Symbolic
             }
         }
 
-        /// <summary>Gets the core state corresponding to the NFA state</summary>
-        public DfaMatchingState<TSet> GetCoreState(int nfaStateId)
+        /// <summary>Gets the core state Id corresponding to the NFA state</summary>
+        public int GetCoreStateId(int nfaStateId)
         {
             Debug.Assert(_stateArray is not null);
             Debug.Assert(nfaStateId < _nfaStateArray.Length);
             Debug.Assert(_nfaStateArray[nfaStateId] < _stateArray.Length);
-            return _stateArray[_nfaStateArray[nfaStateId]];
+            return _nfaStateArray[nfaStateId];
+        }
+
+        /// <summary>Gets the core state corresponding to the NFA state</summary>
+        public DfaMatchingState<TSet> GetCoreState(int nfaStateId)
+        {
+            Debug.Assert(_stateArray is not null);
+            return _stateArray[GetCoreStateId(nfaStateId)];
         }
 
         /// <summary>Critical region for defining a new core transition</summary>
@@ -570,13 +584,13 @@ namespace System.Text.RegularExpressions.Symbolic
         public bool TryCreateNewTransition(
             DfaMatchingState<TSet> sourceState, int mintermId, int offset, bool checkThreshold, [NotNullWhen(true)] out DfaMatchingState<TSet>? nextState)
         {
-            Debug.Assert(_delta is not null);
+            Debug.Assert(_delta is not null && _stateArray is not null);
             lock (this)
             {
                 Debug.Assert(offset < _delta.Length);
 
                 // check if meanwhile delta[offset] has become defined possibly by another thread
-                DfaMatchingState<TSet>? targetState = _delta[offset];
+                DfaMatchingState<TSet>? targetState = _stateArray[_delta[offset]];
                 if (targetState is null)
                 {
                     if (checkThreshold && _stateCache.Count >= SymbolicRegexThresholds.NfaThreshold)
@@ -586,7 +600,7 @@ namespace System.Text.RegularExpressions.Symbolic
                     }
 
                     targetState = sourceState.Next(GetMinterm(mintermId));
-                    Volatile.Write(ref _delta[offset], targetState);
+                    Volatile.Write(ref _delta[offset], targetState.Id);
                 }
 
                 nextState = targetState;
@@ -594,10 +608,25 @@ namespace System.Text.RegularExpressions.Symbolic
             }
         }
 
+        /// <summary>Gets or creates a new DFA transition. This version takes and returns state IDs.</summary>
+        public bool TryCreateNewTransition(
+            int sourceStateId, int mintermId, int offset, bool checkThreshold, out int nextStateId)
+        {
+            Debug.Assert(sourceStateId > 0);
+            Debug.Assert(_stateArray is not null);
+            if (TryCreateNewTransition(_stateArray[sourceStateId], mintermId, offset, checkThreshold, out DfaMatchingState<TSet>? nextState))
+            {
+                nextStateId = nextState.Id;
+                return true;
+            }
+            nextStateId = -1;
+            return false;
+        }
+
         /// <summary>Gets or creates a new NFA transition.</summary>
         public int[] CreateNewNfaTransition(int nfaStateId, int mintermId, int nfaOffset)
         {
-            Debug.Assert(_delta is not null);
+            Debug.Assert(_delta is not null && _stateArray is not null);
             lock (this)
             {
                 Debug.Assert(nfaOffset < _nfaDelta.Length);
@@ -609,7 +638,9 @@ namespace System.Text.RegularExpressions.Symbolic
                     // Create the underlying transition from the core state corresponding to the nfa state
                     DfaMatchingState<TSet> coreState = GetCoreState(nfaStateId);
                     int coreOffset = (coreState.Id << _mintermsLog) | mintermId;
-                    DfaMatchingState<TSet>? coreTarget = _delta[coreOffset] ?? CreateNewTransition(coreState, mintermId, coreOffset);
+                    int coreTargetId = _delta[coreOffset];
+                    DfaMatchingState<TSet>? coreTarget = coreTargetId > 0 ?
+                        _stateArray[coreTargetId] : CreateNewTransition(coreState, mintermId, coreOffset);
 
                     SymbolicRegexNode<TSet> node = coreTarget.Node.Kind == SymbolicRegexNodeKind.DisableBacktrackingSimulation ?
                         coreTarget.Node._left! : coreTarget.Node;