--- a/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/SymbolicRegexMatcher.Dgml.cs
+++ b/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Symbolic/SymbolicRegexMatcher.Dgml.cs
@@ -29,7 +29,7 @@ namespace System.Text.RegularExpressions.Symbolic
             foreach (DfaMatchingState<TSet> state in _builder._stateCache)
             {
                 writer.WriteLine("        <Node Id=\"{0}\" Label=\"{0}\" Category=\"State\" Group=\"Collapsed\" StateInfo=\"{1}\">", state.Id, state.DgmlView);
-                if (state.IsInitialState)
+                if (_builder._stateInfo![state.Id].IsInitial)
                 {
                     writer.WriteLine("            <Category Ref=\"InitialState\" />");
                 }
@@ -143,16 +143,17 @@ namespace System.Text.RegularExpressions.Symbolic
                 foreach (DfaMatchingState<TSet> source in builder._stateCache)
                 {
                     // Get the span of entries in delta that gives the transitions for the different minterms
-                    Span<DfaMatchingState<TSet>?> deltas = builder.GetDeltasFor(source);
+                    Span<int> deltas = builder.GetDeltasFor(source);
                     Span<int[]?> nfaDeltas = builder.GetNfaDeltasFor(source);
                     Debug.Assert(deltas.Length == builder._minterms.Length);
                     for (int i = 0; i < deltas.Length; ++i)
                     {
-                        // null entries are transitions not explored yet, so skip them
-                        if (deltas[i] is DfaMatchingState<TSet> target)
+                        // negative entries are transitions not explored yet, so skip them
+                        int targetId = deltas[i];
+                        if (targetId >= 0)
                         {
                             // Get or create the data for this (source,destination) state ID pair
-                            (int Source, int Target) key = (source.Id, target.Id);
+                            (int Source, int Target) key = (source.Id, targetId);
                             if (!result.TryGetValue(key, out (TSet Rule, List<int> NfaTargets) entry))
                             {
                                 entry = (builder._solver.Empty, new List<int>());