This PR addresses the regressions related to the inner matching loop performance in https://github.com/dotnet/runtime/issues/70022 introduced by https://github.com/dotnet/runtime/pull/69839 and includes further optimizations. The changes in the PR are roughly:

- https://github.com/dotnet/runtime/pull/69839 caused a regression by making fixed length marker evaluation character context aware, which made it more expensive. Fixed length markers were evaluated whenever a nullable state was found in the first phase of the search. This PR pulls the evaluation outside the inner matching loop by recording the state ID instead, which in DFA mode is just an `int` assignment. I've also implemented support for fixed length markers in NFA mode.
- Pulls the decision to try to use `RegexFindOptimizations` out of the inner matching loop by using a `IInitialStateHandler` interface and the "templating" technique with structs implementing the interface as generic type arguments that we already use for DFA and NFA mode.
- Introduces a decision outside the inner matching loop to never evaluate contextual nullability for patterns that don't have any anchors. This also uses the "templating" technique with a `INullabilityHandler` interface.
- Tabulate an array of `(bool IsInitial, bool IsDeadend, bool IsNullable, bool CanBeNullable)` in the `SymbolicRegexBuilder` that can be accessed by state IDs. This reduces pointer chasing involved in looking these up from the `DfaMatchingState` and its related `SymbolicRegexNode`. The existing `IsInitial` field from `DfaMatchingState` is removed for some memory saving.
- Makes `CurrentState` hold the ID (an `int`) of the current DFA state instead of the `DfaMatchingState` instance. This allows cached transitions to be taken in DFA mode without ever dereferencing the current `DfaMatchingState`. Coupled with the above optimization the `DfaMatchingState` is rarely needed in typical patterns.

One optimization that I evaluated, but that didn't decisively help was pulling out the decision to handle the "last end of line" special minterm outside the inner matching loop. I'm guessing the `char == '\n' && pos == input.Length - 1` check is already cheap enough that it isn't really visible.

I compared performance after implementing each optimization to check that each was helping. Here's a total before and after comparison for NonBacktracking on all the relevant benchmarks in dotnet/performance. The baseline here is the situation after the regressions in https://github.com/dotnet/runtime/issues/70022.
summary:
better: 58, geomean: 1.360
worse: 5, geomean: 1.057
total diff: 63

| Slower                                                                                                                                         | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Leipzig.Count(Pattern: "Tom\|Sawyer\|Huckleberry\|Finn", Options: NonBacktracking)       |      1.11 |       4785248.08 |       5325761.46 | several?|
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Leipzig.Count(Pattern: "Tom.{10,25}river\|river.{10,25}Tom", Options: NonBacktracking) |      1.06 |       9628288.00 |      10195706.67 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 12, Options: NonBacktracking)                            |      1.05 |           133.03 |           139.37 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 13, Options: NonBacktracking)                            |      1.04 |           134.98 |           139.74 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_SliceSlice.Count(Options: IgnoreCase, NonBacktracking)                                |      1.03 |    2029181900.00 |    2094702200.00 | several?|

| Faster                                                                                                                                                                                                  | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "[^\\n]*", Options: NonBacktracking)                                                                          |      2.59 |      15286261.76 |       5901962.79 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "(?s).*", Options: NonBacktracking)                                                                           |      2.51 |      10022464.58 |       4000825.81 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Mariomkas.Count(Pattern: "[\\w]+://[^/\\s?#]+[^\\s?#]+(?:\\?[^\\s#]*)?(?:#[^\\s]*)?", Options: NonBacktracking)                                |      2.22 |       6771215.28 |       3045911.11 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "\\b\\w+n\\b", Options: NonBacktracking)                                                                      |      2.17 |      10788114.58 |       4964147.06 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: ".*", Options: NonBacktracking)                                                                               |      1.96 |      12170225.00 |       6225097.67 | bimodal |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "\\w+\\s+Holmes", Options: NonBacktracking)                                                                   |      1.88 |       4073781.25 |       2171156.90 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_SliceSlice.Count(Options: NonBacktracking)                                                                                                     |      1.86 |    1099632850.00 |     590537400.00 | several?|
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "[a-zA-Z]+ing", Options: NonBacktracking)                                                                     |      1.82 |       8657798.33 |       4762550.98 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "Holmes", Options: NonBacktracking)                                                                           |      1.80 |        159116.57 |         88287.89 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "\\s[a-zA-Z]{0,12}ing\\s", Options: NonBacktracking)                                                          |      1.78 |       4577136.36 |       2567926.42 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "the\\s+\\w+", Options: NonBacktracking)                                                                      |      1.76 |       1961271.43 |       1115444.84 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "Holmes.{0,25}Watson\|Watson.{0,25}Holmes", Options: NonBacktracking)                                          |      1.72 |        199215.34 |        115912.40 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "\\w+", Options: NonBacktracking)                                                                             |      1.69 |      23110975.00 |      13655700.00 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "\\w+\\s+Holmes\\s+\\w+", Options: NonBacktracking)                                                           |      1.61 |       3377971.43 |       2104111.40 | bimodal |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "Sher[a-z]+\|Hol[a-z]+", Options: NonBacktracking)                                                             |      1.59 |        204743.55 |        128662.45 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "[a-q][^u-z]{13}x", Options: NonBacktracking)                                                                 |      1.57 |         88111.27 |         56102.58 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "the", Options: NonBacktracking)                                                                              |      1.50 |       1033139.24 |        687533.78 | bimodal |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "Sherlock\\s+Holmes", Options: NonBacktracking)                                                               |      1.47 |         86482.44 |         58707.08 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "Sherlock\|Holmes", Options: NonBacktracking)                                                                  |      1.47 |        177346.23 |        120786.17 | bimodal |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Leipzig.Count(Pattern: ".{2,4}(Tom\|Sawyer\|Huckleberry\|Finn)", Options: NonBacktracking)                                                        |      1.41 |      67257150.00 |      47606437.50 | several?|
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "Sherlock\|Street", Options: NonBacktracking)                                                                  |      1.40 |         76250.69 |         54275.02 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Leipzig.Count(Pattern: ".{0,2}(Tom\|Sawyer\|Huckleberry\|Finn)", Options: NonBacktracking)                                                        |      1.40 |      67197850.00 |      47843350.00 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "zqj", Options: NonBacktracking)                                                                              |      1.40 |         53884.39 |         38434.69 | several?|
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "\\p{Lu}", Options: NonBacktracking)                                                                          |      1.37 |       2540423.68 |       1849647.01 | several?|
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 5, Options: NonBacktracking)                                                                                      |      1.37 |           206.84 |           150.82 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 3, Options: NonBacktracking)                                                                                      |      1.37 |           242.66 |           177.47 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "Sherlock", Options: NonBacktracking)                                                                         |      1.34 |         66087.55 |         49360.96 | bimodal |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "Sherlock\|Holmes\|Watson\|Irene\|Adler\|John\|Baker", Options: NonBacktracking)                                    |      1.33 |       3037911.41 |       2280343.52 | several?|
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "\\p{Ll}", Options: NonBacktracking)                                                                          |      1.33 |      40920612.50 |      30728875.00 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "Sherlock\|Holmes\|Watson", Options: NonBacktracking)                                                           |      1.32 |        211902.74 |        160457.22 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "\\p{L}", Options: NonBacktracking)                                                                           |      1.31 |      40560162.50 |      30977100.00 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "Sherlock Holmes", Options: NonBacktracking)                                                                  |      1.31 |         68187.37 |         52127.82 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "The", Options: NonBacktracking)                                                                              |      1.30 |        130572.40 |        100326.33 | several?|
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 4, Options: NonBacktracking)                                                                                      |      1.29 |           181.44 |           140.99 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "aqj", Options: NonBacktracking)                                                                              |      1.26 |         48508.51 |         38491.90 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "aei", Options: NonBacktracking)                                                                              |      1.26 |         59747.35 |         47496.19 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "(?m)^Sherlock Holmes\|Sherlock Holmes$", Options: NonBacktracking)                                            |      1.25 |         64672.47 |         51940.21 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "(?i)the", Options: NonBacktracking)                                                                          |      1.24 |       1523287.44 |       1232909.66 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 10, Options: NonBacktracking)                                                                                     |      1.13 |           123.15 |           108.83 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 2, Options: NonBacktracking)                                                                                      |      1.12 |           161.55 |           143.72 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 6, Options: NonBacktracking)                                                                                      |      1.11 |           114.45 |           103.11 | several?|
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 7, Options: NonBacktracking)                                                                                      |      1.09 |           101.74 |            93.39 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 9, Options: NonBacktracking)                                                                                      |      1.08 |           112.40 |           103.77 | bimodal |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Mariomkas.Count(Pattern: "(?:(?:25[0-5]\|2[0-4][0-9]\|[01]?[0-9][0-9])\\.){3}(?:25[0-5]\|2[0-4][0-9]\|[01]?[0-9][0-9])", Options: NonBacktracking) |      1.07 |      13209031.58 |      12307950.00 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "(?i)Sher[a-z]+\|Hol[a-z]+", Options: NonBacktracking)                                                         |      1.06 |        984012.25 |        924302.87 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Leipzig.Count(Pattern: "[a-z]shing", Options: NonBacktracking)                                                                                 |      1.06 |       2657380.00 |       2513476.53 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Mariomkas.Count(Pattern: "[\\w\\.+-]+@[\\w\\.-]+\\.[\\w\\.-]+", Options: NonBacktracking)                                                      |      1.06 |        638071.50 |        603662.24 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Leipzig.Count(Pattern: "Twain", Options: NonBacktracking)                                                                                      |      1.05 |       2494026.73 |       2386224.76 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 0, Options: NonBacktracking)                                                                                      |      1.04 |            83.28 |            79.90 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_BoostDocs_Simple.IsMatch(Id: 8, Options: NonBacktracking)                                                                                      |      1.04 |           106.23 |           102.07 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Leipzig.Count(Pattern: "([A-Za-z]awyer\|[A-Za-z]inn)\\s", Options: NonBacktracking)                                                             |      1.04 |      15942657.14 |      15355812.50 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Leipzig.Count(Pattern: "(?i)Tom\|Sawyer\|Huckleberry\|Finn", Options: NonBacktracking)                                                            |      1.04 |      75034250.00 |      72294600.00 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "(?i)Sherlock", Options: NonBacktracking)                                                                     |      1.04 |        103258.57 |         99506.47 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "(?i)Sherlock\|Holmes\|Watson\|Irene\|Adler\|John\|Baker", Options: NonBacktracking)                                |      1.03 |       3304392.00 |       3202143.59 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "(?i)Holmes", Options: NonBacktracking)                                                                       |      1.03 |        507619.96 |        492948.24 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_RustLang_Sherlock.Count(Pattern: "(?i)Sherlock\|Holmes\|Watson", Options: NonBacktracking)                                                       |      1.03 |       2190175.00 |       2130422.17 |         |
| System.Text.RegularExpressions.Tests.Perf_Regex_Industry_Leipzig.Count(Pattern: "(?i)Twain", Options: NonBacktracking)                                                                                  |      1.03 |       6689778.38 |       6510337.84 |         |
…