@EgorBo 2022-06-28
Improvements on x64: https://github.com/dotnet/perf-autofiling-issues/issues/6037